Save file descriptors as int in userspace
Trying to fight against the Unix/Linux default for this just feels pointless
with how much annoyance its caused

+++ include/kanawha/uapi/file.h
@@ -1,7 +1,7 @@
 #ifndef __KANAWHA__UAPI_FILE_H__
 #define __KANAWHA__UAPI_FILE_H__
 
-typedef unsigned long fd_t;
+typedef int fd_t;
 
 // The maximum amount of data which can be transferred
 // by a single read/write system call
